+++ Library/src/main/java/dev/testify/ScreenshotRule.kt
@@ -29,76 +29,46 @@ package dev.testify
 import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Intent
-import android.graphics.Bitmap
 import android.os.Bundle
-import android.os.Debug
-import android.os.Looper
-import android.view.View
-import android.view.ViewGroup
+import android.view.View.NO_ID
 import androidx.annotation.CallSuper
 import androidx.annotation.IdRes
 import androidx.annotation.LayoutRes
-import androidx.annotation.UiThread
 import androidx.annotation.VisibleForTesting
-import androidx.test.annotation.ExperimentalTestApi
 import androidx.test.platform.app.InstrumentationRegistry.getInstrumentation
 import androidx.test.rule.ActivityTestRule
-import dev.testify.annotation.ScreenshotInstrumentation
 import dev.testify.annotation.TestifyLayout
-import dev.testify.internal.DEFAULT_FOLDER_FORMAT
-import dev.testify.internal.DEFAULT_NAME_FORMAT
-import dev.testify.internal.DeviceStringFormatter
+import dev.testify.annotation.findAnnotation
+import dev.testify.annotation.getScreenshotAnnotationName
+import dev.testify.annotation.getScreenshotInstrumentationAnnotation
 import dev.testify.internal.ScreenshotRuleCompatibilityMethods
 import dev.testify.internal.TestifyConfiguration
-import dev.testify.internal.assertExpectedDevice
 import dev.testify.internal.exception.ActivityNotRegisteredException
 import dev.testify.internal.exception.AssertSameMustBeLastException
-import dev.testify.internal.exception.FailedToCaptureBitmapException
-import dev.testify.internal.exception.FinalizeDestinationException
 import dev.testify.internal.exception.MissingAssertSameException
 import dev.testify.internal.exception.MissingScreenshotInstrumentationAnnotationException
-import dev.testify.internal.exception.NoScreenshotsOnUiThreadException
-import dev.testify.internal.exception.ScreenshotBaselineNotDefinedException
-import dev.testify.internal.exception.ScreenshotIsDifferentException
 import dev.testify.internal.exception.ScreenshotTestIgnoredException
-import dev.testify.internal.exception.ViewModificationException
-import dev.testify.internal.extensions.TestInstrumentationRegistry.Companion.getModuleName
-import dev.testify.internal.extensions.TestInstrumentationRegistry.Companion.instrumentationPrintln
-import dev.testify.internal.extensions.cyan
-import dev.testify.internal.extensions.getScreenshotAnnotationName
 import dev.testify.internal.extensions.isInvokedFromPlugin
-import dev.testify.internal.formatDeviceString
 import dev.testify.internal.helpers.ActivityProvider
 import dev.testify.internal.helpers.EspressoActions
 import dev.testify.internal.helpers.EspressoHelper
 import dev.testify.internal.helpers.ResourceWrapper
-import dev.testify.internal.helpers.findRootView
 import dev.testify.internal.helpers.registerActivityProvider
-import dev.testify.internal.logic.compareBitmaps
-import dev.testify.internal.logic.takeScreenshot
-import dev.testify.internal.processor.capture.createBitmapFromDrawingCache
-import dev.testify.internal.processor.diff.HighContrastDiff
-import dev.testify.output.getDestination
+import dev.testify.internal.logic.AssertionState
+import dev.testify.internal.logic.ScreenshotLifecycleHost
+import dev.testify.internal.logic.ScreenshotLifecycleObserver
+import dev.testify.internal.logic.assertSame
 import dev.testify.report.ReportSession
 import dev.testify.report.Reporter
-import org.junit.Assert.assertTrue
-import org.junit.Assume.assumeTrue
 import org.junit.AssumptionViolatedException
 import org.junit.rules.TestRule
 import org.junit.runner.Description
 import org.junit.runners.model.Statement
-import java.util.concurrent.CountDownLatch
-import java.util.concurrent.TimeUnit
-import dev.testify.internal.extensions.TestInstrumentationRegistry.Companion.isRecordMode as recordMode
-
-typealias ViewModification = (rootView: ViewGroup) -> Unit
-typealias ViewProvider = (rootView: ViewGroup) -> View
-typealias ExtrasProvider = (bundle: Bundle) -> Unit
 
 @Suppress("unused", "MemberVisibilityCanBePrivate")
 open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
     protected val activityClass: Class<T>,
-    @IdRes var rootViewId: Int = android.R.id.content,
+    @IdRes override var rootViewId: Int = android.R.id.content,
     initialTouchMode: Boolean = false,
     enableReporter: Boolean = false,
     protected val configuration: TestifyConfiguration = TestifyConfiguration()
@@ -106,6 +76,8 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
     TestRule,
     ActivityProvider<T>,
     ScreenshotLifecycle,
+    AssertionState,
+    ScreenshotLifecycleHost by ScreenshotLifecycleObserver(),
     CompatibilityMethods<ScreenshotRule<T>, T> by ScreenshotRuleCompatibilityMethods() {
 
     @Deprecated(
@@ -126,32 +98,25 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
         configuration = TestifyConfiguration()
     )
 
-    @LayoutRes private var targetLayoutId: Int = NO_ID
+    @LayoutRes override var targetLayoutId: Int = NO_ID
 
     internal val testContext = getInstrumentation().context
-    private var assertSameInvoked = false
+    override var assertSameInvoked = false
     internal val espressoHelper: EspressoHelper by lazy { EspressoHelper(configuration) }
-    private var screenshotViewProvider: ViewProvider? = null
-    private var throwable: Throwable? = null
-    private var viewModification: ViewModification? = null
+    override var screenshotViewProvider: ViewProvider? = null
+    override var throwable: Throwable? = null
+    override var viewModification: ViewModification? = null
     private var extrasProvider: ExtrasProvider? = null
-    private var isRecordMode: Boolean = false
 
     @VisibleForTesting
     internal var reporter: Reporter? = null
         private set
     private lateinit var outputFileName: String
-    private val screenshotLifecycleObservers = HashSet<ScreenshotLifecycle>()
 
     init {
         if (enableReporter || TestifyFeatures.Reporter.isEnabled(getInstrumentation().context)) {
             reporter = Reporter.create(getInstrumentation().targetContext, ReportSession())
         }
-        addScreenshotObserver(TestifyFeatures)
-    }
-
-    private fun isRunningOnUiThread(): Boolean {
-        return Looper.getMainLooper().thread == Thread.currentThread()
     }
 
     fun setRootViewId(@IdRes rootViewId: Int): ScreenshotRule<T> {
@@ -180,21 +145,13 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
         return this
     }
 
-    fun setScreenshotViewProvider(viewProvider: ViewProvider): ScreenshotRule<T> {
+    override fun setScreenshotViewProvider(viewProvider: ViewProvider): ScreenshotRule<T> {
         this.screenshotViewProvider = viewProvider
         return this
     }
 
     fun withExperimentalFeatureEnabled(feature: TestifyFeatures): ScreenshotRule<T> {
         feature.setEnabled(true)
-        return this
-    }
-
-    /**
-     * Record a new baseline when running the test
-     */
-    fun setRecordModeEnabled(isRecordMode: Boolean): ScreenshotRule<T> {
-        this.isRecordMode = isRecordMode
         return this
     }
 
@@ -214,7 +171,7 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
         super.afterActivityLaunched()
         ResourceWrapper.afterActivityLaunched(activity)
         configuration.afterActivityLaunched(activity)
-        screenshotLifecycleObservers.forEach { it.applyConfiguration(activity, configuration) }
+        notifyObservers { it.applyConfiguration(activity, configuration) }
     }
 
     @CallSuper
@@ -234,11 +191,14 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
      * @return a new statement, which may be the same as base, a wrapper around base, or a completely new [Statement].
      */
     override fun apply(base: Statement, description: Description): Statement {
+        addScreenshotObserver(TestifyFeatures)
+
         withRule(this)
 
         val methodAnnotations = description.annotations
         apply(description.methodName, description.testClass, methodAnnotations)
-        statement = ScreenshotStatement(base)
+        val statement = ScreenshotStatement(base)
+        this.statement = statement
         return super.apply(statement, description)
     }
 
@@ -265,6 +225,7 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
         configuration.applyAnnotations(methodAnnotations)
 
         espressoHelper.reset()
+        addScreenshotObserver(espressoHelper)
 
         getInstrumentation().testDescription = TestDescription(
             methodName = methodName,
@@ -272,16 +233,8 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
         )
         reporter?.startTest(getInstrumentation().testDescription)
 
-        val testifyLayout = methodAnnotations?.getAnnotation<TestifyLayout>()
-        targetLayoutId = testifyLayout?.resolvedLayoutId ?: View.NO_ID
-    }
-
-    private inline fun <reified T : Annotation> Collection<Annotation>.getAnnotation(): T? {
-        return this.find { it is T } as? T
-    }
-
-    private inline fun <reified T : Annotation> Collection<Annotation>.getAnnotation(name: String): T? {
-        return this.find { it.annotationClass.qualifiedName == name } as? T
+        val testifyLayout = methodAnnotations?.findAnnotation<TestifyLayout>()
+        targetLayoutId = testifyLayout?.resolvedLayoutId ?: NO_ID
     }
 
     @get:LayoutRes
@@ -294,20 +247,6 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
             }
             return layoutId
         }
-
-    /**
-     * Get the [ScreenshotInstrumentation] instance associated with the test method
-     *
-     * @param classAnnotations - A [List] of all the [Annotation]s defined on the currently running test class
-     * @param methodAnnotations - A [Collection] of all the [Annotation]s defined on the currently running test method
-     */
-    internal fun getScreenshotInstrumentationAnnotation(
-        classAnnotations: List<Annotation>,
-        methodAnnotations: Collection<Annotation>?
-    ): Annotation? {
-        val annotationName = getScreenshotAnnotationName()
-        return classAnnotations.getAnnotation(annotationName) ?: methodAnnotations?.getAnnotation(annotationName)
-    }
 
     /**
      * Assert that the @ScreenshotInstrumentation is defined on the test method.
@@ -366,6 +305,10 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
         return intent
     }
 
+    override fun assureActivity(intent: Intent?) {
+        launchActivity(intent)
+    }
+
     override fun launchActivity(startIntent: Intent?): T {
         try {
             return super.launchActivity(startIntent)
@@ -377,14 +320,6 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
         }
     }
 
-    fun addScreenshotObserver(observer: ScreenshotLifecycle) {
-        this.screenshotLifecycleObservers.add(observer)
-    }
-
-    fun removeScreenshotObserver(observer: ScreenshotLifecycle) {
-        this.screenshotLifecycleObservers.remove(observer)
-    }
-
     /**
      * Test lifecycle method.
      * Invoked immediately before assertSame and before the activity is launched.
@@ -394,181 +329,21 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
         getInstrumentation().registerActivityProvider(this)
     }
 
-    /**
-     * Given [baselineBitmap] and [currentBitmap], use [HighContrastDiff] to write a companion .diff image for the
-     * current test.
-     *
-     * This diff image is a high-contrast image where each difference, regardless of how minor, is indicated in red
-     * against a black background.
-     */
-    @ExperimentalTestApi
-    open fun generateHighContrastDiff(baselineBitmap: Bitmap, currentBitmap: Bitmap) {
-        HighContrastDiff(configuration.exclusionRects)
-            .name(outputFileName)
-            .baseline(baselineBitmap)
-            .current(currentBitmap)
-            .exactness(configuration.exactness)
-            .generate(context = activity)
-    }
-
-    @ExperimentalTestApi
     fun assertSame() {
-        assertSameInvoked = true
         addScreenshotObserver(this)
-        screenshotLifecycleObservers.forEach { it.beforeAssertSame() }
-
-        if (isRunningOnUiThread()) {
-            throw NoScreenshotsOnUiThreadException()
-        }
-
         try {
-            launchActivity(activityIntent)
-        } catch (e: ScreenshotTestIgnoredException) {
-            // Exit gracefully; mark test as ignored
-            assumeTrue(false)
-            return
-        }
-
-        try {
-            try {
-                val description = getInstrumentation().testDescription
-                reporter?.captureOutput(this)
-                outputFileName = formatDeviceString(
-                    DeviceStringFormatter(
-                        testContext,
-                        description.nameComponents
-                    ),
-                    DEFAULT_NAME_FORMAT
-                )
-
-                screenshotLifecycleObservers.forEach { it.beforeInitializeView(activity) }
-                initializeView(activity)
-                screenshotLifecycleObservers.forEach { it.afterInitializeView(activity) }
-
-                espressoHelper.beforeScreenshot()
-
-                val rootView = activity.findRootView(rootViewId)
-                val screenshotView: View? = screenshotViewProvider?.invoke(rootView)
-
-                configuration.beforeScreenshot(rootView)
-
-                screenshotLifecycleObservers.forEach { it.beforeScreenshot(activity) }
-
-                val currentBitmap = takeScreenshot(
-                    activity,
-                    outputFileName,
-                    screenshotView,
-                    configuration.captureMethod ?: ::createBitmapFromDrawingCache
-                ) ?: throw FailedToCaptureBitmapException()
-
-                screenshotLifecycleObservers.forEach { it.afterScreenshot(activity, currentBitmap) }
-
-                if (configuration.pauseForInspection) {
-                    Thread.sleep(LAYOUT_INSPECTION_TIME_MS.toLong())
-                }
-
-                assertExpectedDevice(testContext, description.name, isRecordMode)
-
-                val destination = getDestination(activity, outputFileName)
-
-                val baselineBitmap = loadBaselineBitmapForComparison(testContext, description.name)
-                    ?: if (isRecordMode || recordMode) {
-                        instrumentationPrintln(
-                            "\n\t✓ " + "Recording baseline for ${description.name}".cyan()
-                        )
-
-                        if (!destination.finalize())
-                            throw FinalizeDestinationException(destination.description)
-
-                        return
-                    } else {
-                        throw ScreenshotBaselineNotDefinedException(
-                            moduleName = getModuleName(),
-                            testName = description.name,
-                            testClass = description.fullyQualifiedTestName,
-                            deviceKey = formatDeviceString(
-                                DeviceStringFormatter(
-                                    testContext,
-                                    null
-                                ),
-                                DEFAULT_FOLDER_FORMAT
-                            )
-                        )
-                    }
-
-                if (compareBitmaps(baselineBitmap, currentBitmap, configuration.getBitmapCompare())) {
-                    assertTrue(
-                        "Could not delete cached bitmap ${description.name}",
-                        deleteBitmap(destination)
-                    )
-                } else {
-                    if (!destination.finalize())
-                        throw FinalizeDestinationException(destination.description)
-
-                    if (TestifyFeatures.GenerateDiffs.isEnabled(activity)) {
-                        generateHighContrastDiff(baselineBitmap, currentBitmap)
-                    }
-                    if (isRecordMode || recordMode) {
-                        instrumentationPrintln(
-                            "\n\t✓ " + "Recording baseline for ${description.name}".cyan()
-                        )
-                    } else {
-                        throw ScreenshotIsDifferentException(getModuleName(), description.fullyQualifiedTestName)
-                    }
-                }
-            } finally {
-            }
+            assertSame(
+                state = this,
+                configuration = configuration,
+                testContext = testContext,
+                screenshotLifecycleHost = this,
+                activityProvider = this,
+                activityIntent = activityIntent,
+                reporter = reporter
+            )
         } finally {
-            ResourceWrapper.afterTestFinished(activity)
-            configuration.afterTestFinished()
-            TestifyFeatures.reset()
             removeScreenshotObserver(this)
-            if (throwable != null) {
-                //noinspection ThrowFromfinallyBlock
-                throw RuntimeException(throwable)
-            }
-        }
-    }
-
-    @UiThread
-    @CallSuper
-    open fun applyViewModifications(parentView: ViewGroup) {
-        configuration.applyViewModificationsMainThread(parentView)
-    }
-
-    @VisibleForTesting
-    internal fun initializeView(activity: Activity) {
-        val parentView = activity.findRootView(rootViewId)
-        val latch = CountDownLatch(1)
-
-        var viewModificationException: Throwable? = null
-        activity.runOnUiThread {
-            if (targetLayoutId != NO_ID) {
-                activity.layoutInflater.inflate(targetLayoutId, parentView, true)
-            }
-
-            viewModification?.let { viewModification ->
-                try {
-                    viewModification(parentView)
-                } catch (exception: Throwable) {
-                    viewModificationException = exception
-                }
-            }
-
-            applyViewModifications(parentView)
-
-            latch.countDown()
-        }
-        configuration.applyViewModificationsTestThread(activity)
-
-        if (Debug.isDebuggerConnected()) {
-            latch.await()
-        } else {
-            assertTrue(latch.await(INFLATE_TIMEOUT_SECONDS, TimeUnit.SECONDS))
-        }
-
-        viewModificationException?.let {
-            throw ViewModificationException(it)
+            removeScreenshotObserver(espressoHelper)
         }
     }
 
@@ -621,10 +396,4 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
             field = value
             assertSameInvoked = value
         }
-
-    companion object {
-        const val NO_ID = -1
-        private const val LAYOUT_INSPECTION_TIME_MS = 60000
-        private const val INFLATE_TIMEOUT_SECONDS: Long = 5
-    }
 }