+++ Library/src/main/java/dev/testify/ViewModification.kt
@@ -1,8 +1,8 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2022 ndtp
- *
+ * Copyright (c) 2023 ndtp
+  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
@@ -21,30 +21,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package dev.testify.internal.helpers
+package dev.testify
 
-import android.app.Activity
-import androidx.test.espresso.Espresso
-import dev.testify.ScreenshotLifecycle
-import dev.testify.internal.TestifyConfiguration
+import android.view.ViewGroup
 
-typealias EspressoActions = () -> Unit
-
-class EspressoHelper(private val configuration: TestifyConfiguration) : ScreenshotLifecycle {
-
-    var actions: EspressoActions? = null
-
-    fun reset() {
-        actions = null
-    }
-
-    override fun afterInitializeView(activity: Activity) {
-        actions?.invoke()
-
-        Espresso.onIdle()
-
-        if (configuration.hideSoftKeyboard) {
-            Espresso.closeSoftKeyboard()
-        }
-    }
-}
+typealias ViewModification = (rootView: ViewGroup) -> Unit