+++ Library/src/main/java/dev/testify/internal/logic/AssertionState.kt
@@ -1,9 +1,8 @@
 /*
  * The MIT License (MIT)
  *
- * Modified work copyright (c) 2022 ndtp
- * Original work copyright (c) 2019 Shopify Inc.
- *
+ * Copyright (c) 2023 ndtp
+  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
@@ -22,28 +21,50 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
+package dev.testify.internal.logic
 
-package dev.testify.annotation
-
-import android.view.View.NO_ID
+import androidx.annotation.IdRes
 import androidx.annotation.LayoutRes
+import dev.testify.ViewModification
+import dev.testify.ViewProvider
 
 /**
- * The [TestifyLayout] annotation allows you to specify a layout resource to be automatically
- * loaded into the host Activity for testing.
- *
- * @sample dev.testify.ExperimentalPixelCopyTest.withoutPixelCopy
- *
- * @param layoutId: This integer parameter is a layout resource reference
- *      (e.g. {@code android.R.layout.list_content}).
- * @param layoutResName: A fully qualified resource name of the form "package:type/entry"
- *      (e.g. {@code android:layout/list_content}).
- *      Unfortunately R fields are not constants in Android library projects. LayoutRes IDs cannot
- *      be used as annotations parameters.
+ * This interface defines required internal state for the assertSame() logic.
  */
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY_GETTER, AnnotationTarget.PROPERTY_SETTER)
-annotation class TestifyLayout(
-    @LayoutRes val layoutId: Int = NO_ID,
-    val layoutResName: String = ""
-)
+interface AssertionState {
+
+    /**
+     * Track if the assertSame() method was invoked by the caller.
+     */
+    var assertSameInvoked: Boolean
+
+    /**
+     * Caught exception that thrown during assertSame().
+     */
+    var throwable: Throwable?
+
+    /**
+     * The ID of the [android.view.View] to screenshot.
+     */
+    @get:IdRes var rootViewId: Int
+
+    /**
+     * The ID of the XML layout file to be inflated.
+     */
+    @get:LayoutRes var targetLayoutId: Int
+
+    /**
+     * The [ViewProvider] instance used for taking a screenshot.
+     */
+    val screenshotViewProvider: ViewProvider?
+
+    /**
+     * The [ViewModification] to apply to the activity prior to taking the screenshot.
+     */
+    var viewModification: ViewModification?
+
+    /**
+     * Builder method for setting [screenshotViewProvider]
+     */
+    fun setScreenshotViewProvider(viewProvider: ViewProvider): AssertionState
+}